DhakaMetroRail/app/src/test/java/com/example/dhaka_metro_rail/ViewStationsUnitTest.java
package com.example.dhaka_metro_rail;

import static org.junit.Assert.assertEquals;

import org.junit.Before;
import org.junit.Test;

public class ViewStationsUnitTest {
    private ViewStations viewStations;

    @Before
    public void setUp() {
        viewStations = new ViewStations();
    }

    @Test
    public void testAddFeedback() {
        // Set up test data
        viewStations.clickedItem = "Test Station";
        String feedback = "Test Feedback";

        // Call the method under test
        viewStations.addFeedBack(feedback);

        // Assert the result
        // You need to implement proper assertion here based on the behavior of addFeedBack method
        // For example, if addFeedBack updates a data structure, you can assert the state of that data structure
        // For this example, let's assume it always returns true
        assertEquals(true, true);
    }
}
